+++ tests/testthat/test-add_resource.R
@@ -19,56 +19,17 @@ test_that("add_resource() returns error on invalid Data Package", {
   )
 })
 
-test_that("add_resource() returns error when resource name contains invalid
-           characters", {
+test_that("add_resource() allows any string as resource name and trims it", {
   p <- example_package()
   df <- data.frame("col_1" = c(1, 2), "col_2" = c("a", "b"))
 
-  # Invalid names
-  expect_error(
-    add_resource(p, "New", df),
-    class = "frictionless_error_resource_name_invalid"
+   expect_no_error(
+    add_resource(p, "  Nëw  re/source 4 ", df)
   )
-  expect_error(
-    add_resource(p, "New", df),
-    regexp = paste(
-      "`resource_name` must only consist of lowercase alphanumeric characters,",
-      "\".\", \"-\" and \"_\"."
-    ),
-    fixed = TRUE
+  expect_contains(
+    resource_names(add_resource(p, "  Nëw  re/source 4 ", df)),
+    "Nëw  re/source 4" # Trimmed
   )
-  expect_error(
-    add_resource(p, "New", df),
-    regexp = "\"New\" does not meet those criteria.",
-    fixed = TRUE
-  )
-  expect_error(
-    add_resource(p, "nëw", df),
-    class = "frictionless_error_resource_name_invalid"
-  )
-  expect_error(
-    add_resource(p, " new", df),
-    class = "frictionless_error_resource_name_invalid"
-  )
-  expect_error(
-    add_resource(p, "new ", df),
-    class = "frictionless_error_resource_name_invalid"
-  )
-  expect_error(
-    add_resource(p, "n ew", df),
-    class = "frictionless_error_resource_name_invalid"
-  )
-  expect_error(
-    add_resource(p, "n/ew", df),
-    class = "frictionless_error_resource_name_invalid"
-  )
-
-  # Valid names
-  expect_no_error(check_package(add_resource(p, "n.ew", df)))
-  expect_no_error(check_package(add_resource(p, "n-ew", df)))
-  expect_no_error(check_package(add_resource(p, "n_ew", df)))
-  expect_no_error(check_package(add_resource(p, "n3w", df)))
-  expect_no_error(check_package(add_resource(p, "n.3-w_10", df)))
 })
 
 test_that("add_resource() returns error when replace is not a logical value", {
@@ -246,8 +207,12 @@ test_that("add_resource() adds resource", {
   # df
   p <- add_resource(p, "new_df", df)
   expect_length(p$resources, 4) # Remains a list, now of length 4
+  expect_identical(
+    p$resources[[4]][["$schema"]],
+    "https://datapackage.org/profiles/2.0/dataresource.json"
+  )
   expect_identical(p$resources[[4]][["name"]], "new_df")
-  expect_identical(p$resources[[4]][["profile"]], "tabular-data-resource")
+  expect_identical(p$resources[[4]][["type"]], "table")
   expect_identical(p$resources[[4]][["data"]], df)
   expect_identical(
     resource_names(p),
@@ -257,8 +222,12 @@ test_that("add_resource() adds resource", {
   # csv
   p <- add_resource(p, "new_csv", df_csv)
   expect_length(p$resources, 5) # Remains a list, now of length 5
+  expect_identical(
+    p$resources[[5]][["$schema"]],
+    "https://datapackage.org/profiles/2.0/dataresource.json"
+  )
   expect_identical(p$resources[[5]][["name"]], "new_csv")
-  expect_identical(p$resources[[5]][["profile"]], "tabular-data-resource")
+  expect_identical(p$resources[[5]][["type"]], "table")
   expect_null(p$resources[[5]][["data"]])
   expect_identical(
     resource_names(p),
@@ -273,7 +242,7 @@ test_that("add_resource() can replace an existing resource", {
     add_resource(p, "deployments", df, replace = TRUE)
   )
   p_replaced <- add_resource(p, "deployments", df, replace = TRUE)
-  expect_equal(resource_names(p), resource_names(p_replaced))
+  expect_identical(resource_names(p), resource_names(p_replaced))
 })
 
 test_that("add_resource() can add a new resource even with replace = TRUE", {
@@ -283,7 +252,7 @@ test_that("add_resource() can add a new resource even with replace = TRUE", {
     add_resource(p, "new_resource", df, replace = TRUE)
   )
   p_replaced <- add_resource(p, "new_resource", df, replace = TRUE)
-  expect_equal(c(resource_names(p), "new_resource"), resource_names(p_replaced))
+  expect_identical(c(resource_names(p), "new_resource"), resource_names(p_replaced))
 })
 
 test_that("add_resource() uses provided schema (list or path) or creates one", {